TD-1445: Consolidate model selection service
